% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{The price of x USD in EUR}
\usage{
convert_usd_to_eur(usd, decimals = 4)
}
\arguments{
\item{decimals}{number of decimals for the exchange rate}

\item{x}{number of USD}
}
\value{
integer
}
\description{
The price of x USD in EUR
}
\examples{
convert_usd_to_eur(10)
convert_usd_to_eur(100, 2)
}
