% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\name{eur}
\alias{eur}
\title{Value in € format}
\usage{
eur(amount)
}
\arguments{
\item{amount}{amount to transform to €}
}
\value{
string
}
\description{
Value in € format
}
\examples{
eur(10000)
eur(1234.5678)
}
\seealso{
uneur()
}
