% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{The price of 1 USD in EUR}
\usage{
eurusd(decimals = 4)
}
\arguments{
\item{decimals}{number of decimals to show}
}
\value{
integer
}
\description{
The price of 1 USD in EUR
}
\examples{
eurusd()
eurusd(2)
}
