% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{Historical prices of 1 USD in EUR}
\usage{
eurusds(date_from, date_to, decimals = 4)
}
\arguments{
\item{date_from}{starting date of the period, back-filled for missing dates}

\item{date_to}{ending date of the period, back-filled for missing dates}

\item{decimals}{number of decimals to show}
}
\value{
integer
}
\description{
Historical prices of 1 USD in EUR
}
\examples{
eurusds('2018-01-01', '2018-01-10')
eurusds('2018-01-01', '2018-01-10', 2)
}
