% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{Value stripped of € currency sign}
\usage{
uneur(eur)
}
\arguments{
\item{eur}{€ value}
}
\value{
integer
}
\description{
Value stripped of € currency sign
}
\examples{
uneur(eur(10000))
uneur('€123,456.789')
uneur('-€123,456.789')
}
\seealso{
eur()
}
